user_management/requirements.txt
marshmallow
sqlalchemy
wheel
flask
python-abc
python-dotenv
psycopg2-binary
requests
exceptiongroup
pytest
pytest-cov
pytest-mock
pytest-env
bcrypt